S3Client on root infra folder
