+++ app/api/files/storage.ts
@@ -1,5 +1,4 @@
-import { NoSuchKey, NotFound, S3Client } from '@aws-sdk/client-s3';
-import { NodeHttpHandler, NodeHttpHandlerOptions } from '@smithy/node-http-handler';
+import { NoSuchKey, NotFound } from '@aws-sdk/client-s3';
 import { inspect } from 'util';
 // eslint-disable-next-line node/no-restricted-import
 import { createReadStream, createWriteStream } from 'fs';
@@ -14,7 +13,7 @@ import { FileType } from 'shared/types/fileType';
 import { Readable } from 'stream';
 import { pipeline } from 'stream/promises';
 
-import { LoggerFactory } from 'api/core/infrastructure/factories/LoggerFactory';
+import { buildS3Client } from 'api/infrastructure/S3Client';
 import { FileNotFound } from './FileNotFound';
 import {
   activityLogPath,
@@ -28,79 +27,9 @@ import { S3Error, S3Storage } from './S3Storage';
 
 let s3Instance: S3Storage;
 
-const buildS3Client = (params: {}) => {
-  const client = new S3Client({
-    maxAttempts: 5,
-    requestHandler: new NodeHttpHandler(params),
-    apiVersion: 'latest',
-    region: 'placeholder-region',
-    endpoint: config.s3.endpoint,
-    credentials: config.s3.credentials,
-    forcePathStyle: true,
-  });
-
-  // eslint-disable-next-line max-statements
-  client.middlewareStack.add((next, context) => async args => {
-    const startTime = Date.now();
-
-    const input = args.input as { Body?: Buffer; Key?: string };
-
-    try {
-      const result = await next(args);
-      const duration = Date.now() - startTime;
-
-      if (process.env.NODE_ENV !== 'test') {
-        LoggerFactory.default().info('S3 operation completed', {
-          operation: context.commandName,
-          duration,
-          key: input.Key,
-          fileSizeKB: Buffer.isBuffer(input.Body) ? input.Body.length / 1024 : NaN,
-          success: true,
-          timestamp: new Date().toISOString(),
-        });
-      }
-
-      return result;
-    } catch (error) {
-      if (process.env.NODE_ENV !== 'test') {
-        LoggerFactory.default().info('S3 operation failed', {
-          operation: context.commandName,
-          duration: Date.now() - startTime,
-          key: input.Key,
-          fileSizeKB: Buffer.isBuffer(input.Body) ? input.Body.length / 1024 : NaN,
-          success: false,
-          error: error.message,
-          errorCode: error.$metadata?.httpStatusCode,
-          timestamp: new Date().toISOString(),
-        });
-      }
-      throw error;
-    }
-  });
-  return client;
-};
 const s3 = () => {
-  const params: NodeHttpHandlerOptions = {
-    socketTimeout: 30000,
-    connectionTimeout: 3000,
-    httpAgent: {
-      maxSockets: 500,
-      timeout: 60000,
-      maxFreeSockets: 100,
-      keepAlive: true,
-      keepAliveMsecs: 5000,
-    },
-    httpsAgent: {
-      maxSockets: 500,
-      timeout: 60000,
-      maxFreeSockets: 100,
-      keepAlive: true,
-      keepAliveMsecs: 5000,
-    },
-  };
-
   if (config.s3.endpoint && !s3Instance) {
-    s3Instance = new S3Storage(buildS3Client(params));
+    s3Instance = new S3Storage(buildS3Client());
   }
 
   return s3Instance;
